Extract themes into a separate module
