+++ app/build.gradle.kts
@@ -106,6 +106,7 @@ android {
 }
 
 dependencies {
+    implementation(projects.core.designsystem)
     implementation(projects.core.navigation)
 
     // Android core
