+++ app/src/main/kotlin/com/oussamateyib/thoth/feature/notes/impl/presentation/editor/util/NoteColors.kt
@@ -1,11 +1,11 @@
 package com.oussamateyib.thoth.feature.notes.impl.presentation.editor.util
 
-import com.oussamateyib.thoth.ui.theme.BabyBlue
-import com.oussamateyib.thoth.ui.theme.LightGreen
-import com.oussamateyib.thoth.ui.theme.RedOrange
-import com.oussamateyib.thoth.ui.theme.RedPink
-import com.oussamateyib.thoth.ui.theme.Violet
-import com.oussamateyib.thoth.ui.theme.White
+import com.oussamateyib.thoth.core.designsystem.theme.BabyBlue
+import com.oussamateyib.thoth.core.designsystem.theme.LightGreen
+import com.oussamateyib.thoth.core.designsystem.theme.RedOrange
+import com.oussamateyib.thoth.core.designsystem.theme.RedPink
+import com.oussamateyib.thoth.core.designsystem.theme.Violet
+import com.oussamateyib.thoth.core.designsystem.theme.White
 
 object NoteColors {
     val noteColors = listOf(RedPink, RedOrange, LightGreen, Violet, BabyBlue, White)
