+++ core/designsystem/src/main/kotlin/com/oussamateyib/thoth/core/designsystem/theme/Color.kt
@@ -1,4 +1,4 @@
-package com.oussamateyib.thoth.ui.theme
+package com.oussamateyib.thoth.core.designsystem.theme
 
 import androidx.compose.ui.graphics.Color
