+++ core/designsystem/src/main/kotlin/com/oussamateyib/thoth/core/designsystem/theme/Type.kt
@@ -1,4 +1,4 @@
-package com.oussamateyib.thoth.ui.theme
+package com.oussamateyib.thoth.core.designsystem.theme
 
 import androidx.compose.material3.Typography
 import androidx.compose.ui.text.TextStyle
@@ -6,7 +6,6 @@ import androidx.compose.ui.text.font.FontFamily
 import androidx.compose.ui.text.font.FontWeight
 import androidx.compose.ui.unit.sp
 
-// Set of Material typography styles to start with
 val Typography = Typography(
     bodyLarge = TextStyle(
         fontFamily = FontFamily.Default,
